API-Python/requirements.txt
Flask==2.0.3
Flask_Cors==3.0.10
Pillow==9.3.0
python-dotenv==0.21.0
stability_sdk==0.3.0
youtube_dl==2021.12.17
